refactor(renderer): #4320 캡션 높이 산식을 composer::caption_height_px 로 통일
렌더(picture_footnote.rs::calculate_caption_height)와 측정
(height_measurer.rs::measure_caption)이 같은 캡션 높이를 각자 재구현하며 갈라져
있었다 — 렌더 쪽만 저장 line_segs 높이와 compose_paragraph 재계산 높이의 max 를
취했고, 측정 쪽은 compose 폴백 없이 line_segs 합산만 했다.

- composer::caption_height_px(caption, dpi) 신설 — 렌더 쪽 산식을 그대로 옮긴 것으로
  새 로직 없음. line_segs 를 하한으로 쓰는 이유: 2d973021c 가
  samples/rowbreak-problem-pages.hwpx 오버랩(issue_rowbreak_chart_overlap.rs)을
  고친 근거가 "폰트 대체 등으로 compose 재계산이 실제 한컴 레이아웃보다 낮게 나올
  수 있다"였다 — 저장값이 한컴 실측이므로 하한 자격이 있다. 측정 쪽엔 이 하한이
  없어 컴포즈 폴백조차 없는 상태였으므로 렌더 쪽을 authority 로 채택.
- calculate_caption_height, measure_caption 은 공유 함수 호출로 교체. 기본 줄 높이
  400 HWPUNIT 상수도 caption_height_px 안 한 곳으로 통합.

검증: issue_rowbreak_chart_overlap(20/20, 오버랩 회귀 포함) + 캡션 관련 tests/
8개 파일(12/12) + lib caption 테스트(63/63) + lineseg_compare(9/9) + fmt/clippy(-D
warnings) 통과.

Closes #4320
